requirements.txt
aiogram==3.7.0
openai==0.27.8
python-dotenv==1.0.1
